feat: add deleteSpinner

+++ src/components/NoteCard/NoteCard.jsx
@@ -3,6 +3,7 @@ import { useState } from "react";
 import CloseRoundedIcon from "@mui/icons-material/CloseRounded";
 import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined";
 import { supabase } from "../../client.js";
+import DeleteSpinner from "../DeleteSpinner/DeleteSpinner.jsx";
 
 export default function NoteCard({
   id,
@@ -13,39 +14,13 @@ export default function NoteCard({
   open,
   clickHandler,
   fetchData,
+  isDeleting,
+  setIsDeleting,
 }) {
   // Create a new Date object
   const [center, setCenter] = useState({ x: -50, y: -50 });
-  const [isEditMode, setIsEditMode] = useState(false); // Edit mode state
   const [editableTitle, setEditableTitle] = useState(title); // Editable title state
   const [editableContent, setEditableContent] = useState(content); // Editable content state
-
-  const toggleEditMode = () => {
-    setIsEditMode(!isEditMode);
-  };
-
-  const handleTitleChange = (e) => {
-    setEditableTitle(e.target.value);
-  };
-
-  const handleContentChange = (e) => {
-    setEditableContent(e.target.value);
-  };
-
-  const updateNotes = async () => {
-    // Save the changes to the database or state
-    const { error } = await supabase
-      .from("notes")
-      .update({ title: editableTitle, content: editableContent })
-      .eq("id", id);
-    if (error) {
-      console.error("Error updating note:", error);
-    } else {
-      console.log("Note updated successfully");
-      fetchData();
-      toggleEditMode(); // Exit edit mode after saving
-    }
-  };
 
   const handleMouseClick = (e) => {
     const xPercentage = -100 + (e.clientX / window.innerWidth) * 100;
@@ -94,6 +69,15 @@ export default function NoteCard({
 
   async function deleteHandler(event) {
     event.stopPropagation();
+
+    const isConfirmed = window.confirm(
+      "Are you sure you want to delete this note?"
+    );
+    if (!isConfirmed) {
+      return; // If the user cancels, exit the function without deleting
+    }
+
+    setIsDeleting(true);
     const { error } = await supabase.from("notes").delete().eq("id", id);
 
     if (error) {
@@ -104,88 +88,117 @@ export default function NoteCard({
     }
   }
 
+  const handleTitleChange = (e) => {
+    setEditableTitle(e.target.value);
+  };
+
+  const handleContentChange = (e) => {
+    setEditableContent(e.target.value);
+  };
+
+  const updateNotes = async () => {
+    // Save the changes to the database or state
+    const { error } = await supabase
+      .from("notes")
+      .update({ title: editableTitle, content: editableContent })
+      .eq("id", id);
+    if (error) {
+      console.error("Error updating note:", error);
+    } else {
+      console.log("Note updated successfully");
+      fetchData();
+    }
+  };
   return (
     <>
-      <div
-        className="note-container"
-        style={{ backgroundColor: colors[priority].primary }}
-        onClick={(e) => {
-          clickHandler(id);
-          handleMouseClick(e); // Attach the handleMouseClick here
-        }}
-      >
-        <>
-          <DeleteOutlineOutlinedIcon
-            className="note-deleteIcon"
-            onClick={deleteHandler}
-          />
-          <div
-            className="note-day"
-            style={{ backgroundColor: colors[priority].secondary }}
-          >
-            {day}
-          </div>
-          <h1 className="note-title">{editableTitle}</h1>
-          <p className="note-content">{editableContent}</p>
-          <div
-            className="note-date"
-            style={{ color: colors[priority].secondary }}
-          >
-            <strong>{dateString}</strong>
-          </div>
-          {/* Insert the date string here */}
-        </>
-      </div>
-      {open && (
+      {isDeleting ? (
+        <DeleteSpinner />
+      ) : (
         <>
           <div
-            className="overlay"
-            onClick={() => {
+            className="note-container"
+            style={{ backgroundColor: colors[priority].primary }}
+            onClick={(e) => {
               clickHandler(id);
-              updateNotes();
-            }}
-          ></div>
-          <div
-            className={`note-container__active ${open ? "opening" : "closing"}`}
-            style={{
-              backgroundColor: colors[priority].primary,
-              "--x": `${center.x}%`,
-              "--y": `${center.y}%`,
+              handleMouseClick(e); // Attach the handleMouseClick here
             }}
           >
-            <div
-              className="note-day"
-              style={{ backgroundColor: colors[priority].secondary }}
-            >
-              {day}
-            </div>
-
             <>
-              <input
-                className="note-title__active"
-                value={editableTitle}
-                onChange={handleTitleChange}
+              <DeleteOutlineOutlinedIcon
+                className="note-deleteIcon"
+                onClick={deleteHandler}
               />
-              <textarea
-                className="note-content__active"
-                value={editableContent}
-                onChange={handleContentChange}
-              />
+              <div
+                className="note-day"
+                style={{ backgroundColor: colors[priority].secondary }}
+              >
+                {day}
+              </div>
+              <h1 className="note-title">{editableTitle}</h1>
+              <p className="note-content">{editableContent}</p>
+              <div
+                className="note-date"
+                style={{ color: colors[priority].secondary }}
+              >
+                <strong>{dateString}</strong>
+              </div>
+              {/* Insert the date string here */}
             </>
-            <div
-              className="note-date"
-              style={{ color: colors[priority].secondary }}
-            >
-              <strong>{dateString}</strong>
-            </div>
-            <CloseRoundedIcon
-              className="overlay-close"
-              onClick={() => {
-                clickHandler(id);
-                updateNotes();
-              }}
-            />
           </div>
+          {open && (
+            <>
+              <div
+                className="overlay"
+                onClick={() => {
+                  clickHandler(id);
+                  updateNotes();
+                }}
+              ></div>
+              <div
+                className={`note-container__active ${
+                  open ? "opening" : "closing"
+                }`}
+                style={{
+                  backgroundColor: colors[priority].primary,
+                  "--x": `${center.x}%`,
+                  "--y": `${center.y}%`,
+                }}
+              >
+                <div
+                  className="note-day"
+                  style={{ backgroundColor: colors[priority].secondary }}
+                >
+                  {day}
+                </div>
+
+                <>
+                  <input
+                    className="note-title__active"
+                    value={editableTitle}
+                    onChange={handleTitleChange}
+                  />
+                  <textarea
+                    className="note-content__active"
+                    value={editableContent}
+                    onChange={handleContentChange}
+                  />
+                </>
+                <div
+                  className="note-date"
+                  style={{ color: colors[priority].secondary }}
+                >
+                  <strong>{dateString}</strong>
+                </div>
+                <CloseRoundedIcon
+                  className="overlay-close"
+                  onClick={() => {
+                    clickHandler(id);
+                    updateNotes();
+                  }}
+                />
+              </div>
+            </>
+          )}
         </>
       )}
     </>
